+++ internal/service/gateway/gateway.go
@@ -25,6 +25,9 @@ const (
 	routerIDDescription         = "ID of the router attached to the gateway."
 	configuredStatusDescription = "The service configured status indicates the service's current intended status. Managed by the customer."
 	operationalStateDescription = "The service operational state indicates the service's current operational, effective state. Managed by the system."
+	addressesDescription        = "IP addresses assigned to the gateway."
+
+	cleanupWaitTimeSeconds = 15
 )
 
 func ResourceGateway() *schema.Resource {
@@ -90,6 +93,25 @@ func ResourceGateway() *schema.Resource {
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
+			"addresses": {
+				Description: addressesDescription,
+				Computed:    true,
+				Type:        schema.TypeSet,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"address": {
+							Type:        schema.TypeString,
+							Description: "IP addresss",
+							Computed:    true,
+						},
+						"name": {
+							Type:        schema.TypeString,
+							Description: "Name of the IP address",
+							Computed:    true,
+						},
+					},
+				},
+			},
 		},
 	}
 }
@@ -120,7 +142,7 @@ func resourceGatewayCreate(ctx context.Context, d *schema.ResourceData, meta int
 
 	d.SetId(gw.UUID)
 
-	err = waitForGatewayToBeRunning(ctx, svc, gw.UUID)
+	gw, err = waitForGatewayToBeRunning(ctx, svc, gw.UUID)
 	if err != nil {
 		return diag.FromErr(err)
 	}
@@ -181,6 +203,9 @@ func resourceGatewayDelete(ctx context.Context, d *schema.ResourceData, meta int
 	// wait before continuing so that router can be deleted if needed
 	diags := diag.FromErr(waitForGatewayToBeDeleted(ctx, svc, d.Id()))
 
+	// Additionally wait some time so that all cleanup operations can finish
+	time.Sleep(time.Second * cleanupWaitTimeSeconds)
+
 	return diags
 }
 
@@ -213,23 +238,35 @@ func setGatewayResourceData(d *schema.ResourceData, gw *upcloud.Gateway) (diags 
 		return diag.FromErr(err)
 	}
 
+	var addresses []map[string]interface{}
+	for _, address := range gw.Addresses {
+		addresses = append(addresses, map[string]interface{}{
+			"address": address.Address,
+			"name":    address.Name,
+		})
+	}
+
+	if err := d.Set("addresses", addresses); err != nil {
+		return diag.FromErr(err)
+	}
+
 	return diags
 }
 
-func waitForGatewayToBeRunning(ctx context.Context, svc *service.Service, id string) error {
-	const maxRetries int = 100
+func waitForGatewayToBeRunning(ctx context.Context, svc *service.Service, id string) (*upcloud.Gateway, error) {
+	const maxRetries int = 500
 
 	for i := 0; i <= maxRetries; i++ {
 		select {
 		case <-ctx.Done():
-			return ctx.Err()
+			return nil, ctx.Err()
 		default:
 			gw, err := svc.GetGateway(ctx, &request.GetGatewayRequest{UUID: id})
 			if err != nil {
-				return err
+				return nil, err
 			}
 			if gw.OperationalState == upcloud.GatewayOperationalStateRunning {
-				return nil
+				return gw, nil
 			}
 
 			tflog.Info(ctx, "waiting for network gateway to be running", map[string]interface{}{"name": gw.Name, "state": gw.OperationalState})
@@ -237,11 +274,11 @@ func waitForGatewayToBeRunning(ctx context.Context, svc *service.Service, id str
 		time.Sleep(5 * time.Second)
 	}
 
-	return fmt.Errorf("max retries (%d)reached while waiting for network gateway to be running", maxRetries)
+	return nil, fmt.Errorf("max retries (%d)reached while waiting for network gateway to be running", maxRetries)
 }
 
 func waitForGatewayToBeDeleted(ctx context.Context, svc *service.Service, id string) error {
-	const maxRetries int = 100
+	const maxRetries int = 500
 
 	for i := 0; i <= maxRetries; i++ {
 		select {